feat(nc-gui): Moved repeated logic of erd to ErdView component and renaned old ErdView as ErdSimpleView
